Moving games sorting into gamesList component
